feat:router autententication and add  @fastfy/jwt

+++ SERVER/src/routes/memories.ts
@@ -1,10 +1,18 @@
 import { FastifyInstance } from "fastify";
 import { z } from 'zod'
 import { prisma } from "./lib/prisma";
+import { request } from "http";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-    app.get('/memories', async () => {
+    app.addHook('preHandler', async (request) => {
+        await request.jwtVerify()
+    })
+    app.get('/memories', async (request) => {
+
         const memories = await prisma.memory.findMany({
+            where: {
+                userId: request.user.sub
+            },
             orderBy: {
                 createdAt: 'asc'
             }
@@ -18,7 +26,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         })
 
     })
-    app.get('/memories/:id', async (request) => {
+    app.get('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
@@ -29,8 +37,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
             }
 
         })
+        if (!memory.isPublic && memory.userId != request.user.sub) {
+            return reply.status(401).send('User not Autorization')
+        }
         return memory
-
 
     })
     app.post('/memories', async (request) => {
@@ -45,13 +55,13 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 content,
                 coverUrl,
                 isPublic,
-                userId: ''
+                userId: request.user.sub
             }
         })
         return memory
 
     })
-    app.put('/memories/:id', async (request) => {
+    app.put('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
@@ -62,7 +72,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
             isPublic: z.coerce.boolean().default(false),
         })
         const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
-        const memory = await prisma.memory.update({
+        let memory = await prisma.memory.findFirstOrThrow({
+            where: {
+                id,
+            }
+        })
+        if (memory.userId != request.user.sub) {
+            return reply.status(401).send('User not Autorization')
+        }
+        memory = await prisma.memory.update({
             where: {
                 id: id,
             },
@@ -75,11 +93,19 @@ export async function memoriesRoutes(app: FastifyInstance) {
         return memory
 
     })
-    app.delete('/memories/:id', async (request) => {
+    app.delete('/memories/:id', async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         })
         const { id } = paramsSchema.parse(request.params)
+        const memory = await prisma.memory.findFirstOrThrow({
+            where: {
+                id,
+            }
+        })
+        if (memory.userId != request.user.sub) {
+            return reply.status(401).send('User not Autorization')
+        }
         await prisma.memory.delete({
             where: {
                 id: id,
